V 0.1

+++ main.py
@@ -147,6 +147,8 @@ async def on_ready():
       main_logger.debug("Mapped MEMBER schedule to role {0}".format(server.Member))
       server.mapRole(discord.utils.find(lambda r: r.name == "Pug",guild.roles),"PUG")
       main_logger.debug("Mapped PUG schedule to role {0}".format(server.PUG))
+      server.mapRole(discord.utils.find(lambda r: r.name == "Schedulizer Admin", guild.roles), "Admin")
+      main_logger.debug("Mapped ADMIN attribute to role {0}".format(server.Admin))
       main_logger.info("Completed processing roles for server {0}".format(server.name))
       # Create user objects for each player with a valid schedule role
       main_logger.info("Processing members for {0}...".format(server.name))
@@ -215,7 +217,7 @@ async def show(context,opt: str):
     events = []
     for key in server.events_db.getall():
       e = server.events_db.get(key)
-      events.append("Type: "+e[0]+"\tName: "+e[2]+"\tDate: "+e[1])
+      events.append("Type: {0}\tName: {1}\tDate: {2}\tOwner: {3}".format(e[0],e[3],e[2],e[1]))
     if len(events) > 0:
       await context.send("\n".join(events))
     else:
@@ -389,11 +391,11 @@ async def event(context, *args):
   if operation == "create":
     if type == 'event':
       if not frequency:
-        e = Event(d,name,rec)
+        e = Event(context.author.name,d,name,rec)
       else:
-        e = Event(d,name,rec,frequency)
+        e = Event(context.author.name,d,name,rec,frequency)
     elif type == 'raid':
-      e = Raid(d,name,rec,frequency)
+      e = Raid(context.author.name,d,name,rec,frequency)
     result = server.addEvent(e)
     message = ""
     if result:
@@ -402,7 +404,22 @@ async def event(context, *args):
     else:
       message = "ERROR: Unable to add event"
       main_logger.error("Server: {0}; Unable to create event".format(context.guild.name))
+  ## DELETE
   elif operation == "delete":
+    ## check permissions
+    ### grab all possible events this could apply to
+    server.event_lock.acquire()
+    es = []
+    for key in server.events_db.getall():
+      e = server.events_db.get(key)
+      ## only track events that the author owns
+      if e[1] == context.author.name:
+        es.append(e)
+    server.event_lock.release()
+    ## if we found no events that the author owns AND the author is not an admin
+    if len(es) < 1 and server.Admin not in context.author.roles:
+      main_logger.error("Server: {0}; Message: Invalid permissions; User {1} attempted to change property {2} of user {3}".format(server.name,context.author,property,player.name))
+      raise discord.ext.commands.CheckFailure("Insufficient permissions; must have the \'Schedulizer Admin\' Role to delete events you do not own")
     result = server.deleteEvent(name,type,date,time)
     if result:
       message = "Successfully deleted {0} instance(s) of event {1}".format(type,name)
@@ -410,7 +427,24 @@ async def event(context, *args):
     else:
       message = "ERROR: Unable to delete event"
       main_logger.error("Server: {0}; Unable to delete event".format(context.guild.name))
+  ## SET
   elif operation == "set":
+    ## check permissions
+    ### grab all possible events this could apply to
+    server.event_lock.acquire()
+    es = []
+    for key in server.events_db.getall():
+      e = server.events_db.get(key)
+      ## only track events that the author owns
+      if e[1] == context.author.name:
+        es.append(e)
+    server.event_lock.release()
+    print(len(es))
+    print(context.author.roles)
+    ## if we found no events that the author owns AND the author is not an admin
+    if len(es) < 1 and server.Admin not in context.author.roles:
+      main_logger.error("Server: {0}; Message: Invalid permissions; User {1} attempted to change property {2} of user {3}".format(server.name,context.author,property,player.name))
+      raise discord.ext.commands.CheckFailure("Insufficient permissions; must have the \'Schedulizer Admin\' Role to modify attributes for events you do not own")
     if property == "date":
       value = datetime.datetime.strptime(value,"%d%b%Y")
     elif property == "time":
@@ -457,6 +491,10 @@ async def player(context, *args):
   player = next(filter(lambda p: p.name == name,server.roster))
   if player == None:
     raise ValueError("Unknown Player")
+  # check permissions
+  if context.author.name != player.name and server.Admin not in context.author.roles:
+    main_logger.error("Server: {0}; Message: Invalid permissions; User {1} attempted to change property {2} of user {3}".format(server.name,context.author,property,player.name))
+    raise discord.ext.commands.CheckFailure("Insufficient permissions; must have the \'Schedulizer Admin\' Role to modify other users' attributes")
   # make sure the player object has a property associated with the input
   if property != None and not hasattr(player,property):
     raise ValueError("Unknown property")
@@ -491,8 +529,8 @@ async def update(context,component):
 @client.event
 async def on_command_error(context,error):
   if isinstance(error,discord.ext.commands.errors.CheckFailure):
-    main_logger.warning("User {0}/{1} attempted to issue command {2}".format(context.user,context.guild.name,context.message))
-    await context.send('You do not have the correct role for this command.')
+    main_logger.warning("User {0}/{1} attempted to issue command {2}".format(context.author,context.guild.name,context.message.content))
+    await context.send(error)
   elif isinstance(error,discord.ext.commands.errors.MissingRequiredArgument):
     await context.send_help(context.command)
   elif isinstance(error,discord.ext.commands.errors.CommandNotFound):
